ucstricmp: compare null and empty strings equal
For consistency with ucstrcmp(), which does the same.

Before, the two overloads would sort null before empty strings. This
behavior was removed in Qt 3.0, IIRC-

This does not yet change anything, as all callers seem to work around
the problem by handling null strings before calling this function
(directly or indirectly). We would have seen a failure crop up if it
wasn't so.

As soon as we use these functions to compare QStringViews, however,
the functions need to deal correctly with a nullptr lhs and/or rhs, so
fix them.

Change-Id: Ie4e417aade993213169b96b5e7351850c52ae733
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>